Decomposed process.ts, Audited and fixed 3 small pre-existing bugs
